+++ packages/esm-patient-chart-app/src/clinical-views/components/encounter-list-tabs.scss
