+++ packages/esm-patient-chart-app/src/index.ts
@@ -27,7 +27,7 @@ import startVisitActionButtonOnPatientSearch from './visit/start-visit-button.co
 import startVisitFormComponent from './visit/visit-form/visit-form.component';
 import stopVisitActionButtonComponent from './actions-buttons/stop-visit.component';
 import visitAttributeTagsComponent from './patient-banner-tags/visit-attribute-tags.component';
-import encounterListTabsComponent from './encounter-list/encounter-list-tabs.component';
+import encounterListTabsComponent from './clinical-views/components/encounter-list-tabs.component';
 
 // This allows @openmrs/esm-patient-common-lib to be accessed by modules that are not
 // using webpack. This is used for ngx-formentry.
